Removing the bad hack intending to avoid "our".
The reluctance against "our" has no valid ground. For reasons associated
with Unicode, the minimum version for the module is 5.8.8. If you do not
use the "date_string" method, you can use the module with an older version,
5.005 or 5.004 (not tested). But the scripts "prt_cal", "g2r_table" and "r2g_table"
rely on "date_string", so they need perl 5.8.8. And "our" is available
with this version.